// src/core/enums/http-statuses.enum.ts
/**
 * Перечисление стандартных HTTP-статусов
 *
 * Используется для упрощённого и безопасного указания кодов ответа
 */
export enum HttpStatus {
    /** Успешный запрос */
    OK = 200,

    /** Ресурс успешно создан */
    Created = 201,

    /** Запрос обработан, но тело ответа отсутствует */
    NoContent = 204,

    /** Некорректные данные в запросе */
    BadRequest = 400,

    /** Ошибка авторизации (пользователь не аутентифицирован) */
    Unauthorized = 401,

    /** Доступ запрещён (нет прав для выполнения операции) */
    Forbidden = 403,

    /** Ресурс не найден */
    NotFound = 404,

    /** Конфликт при обработке запроса */
    Conflict = 409,

    /**Внутренняя ошибка сервера */
    InternalServerError = 500,
}